% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_functions.R
\name{fwrite_tsv}
\alias{fwrite_tsv}
\title{Write data.table to file with preferred settings}
\usage{
fwrite_tsv(x, file)
}
\arguments{
\item{x}{a data.table object}

\item{file}{output file path}
}
\description{
Write data.table to file with preferred settings
}
