% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_functions.R
\name{getQuantile}
\alias{getQuantile}
\title{Get the row from a quantile table to which a value belongs}
\usage{
getQuantile(val, qTable)
}
\arguments{
\item{val}{numeric value}

\item{qTable}{data.frame or data.table with a column "end"}
}
\value{
integer row number
}
\description{
Get the row from a quantile table to which a value belongs
}
