% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_functions.R
\name{myfwrite}
\alias{myfwrite}
\title{Write data.table to file with preferred settings}
\usage{
myfwrite(x, file)
}
\arguments{
\item{x}{data.table}

\item{file}{output file path}
}
\description{
Write data.table to file with preferred settings
}
