% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_functions.R
\name{normaliseQuantile_betweenChr}
\alias{normaliseQuantile_betweenChr}
\title{Quantile normalize count data between chromosomes}
\usage{
normaliseQuantile_betweenChr(
  feature_dat,
  chrs = paste0("chr", c(1:22, "X", "Y")),
  value_col
)
}
\arguments{
\item{feature_dat}{data.table with columns fid and counts}

\item{chrs}{vector of chromosomes to normalise between}

\item{value_col}{name of the column in feature_dat that contains the values to normalize}
}
\value{
data.table with normalised counts
}
\description{
Quantile normalize count data between chromosomes
}
