% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_functions.R
\name{normaliseQuantile_betweenFeature}
\alias{normaliseQuantile_betweenFeature}
\title{Quantile normalize count data between features}
\usage{
normaliseQuantile_betweenFeature(feature_dat, features, qnorm_bchr = FALSE)
}
\arguments{
\item{feature_dat}{data.table with columns fid and counts}

\item{features}{vector of feature names to normalise between}

\item{qnorm_bchr}{logical, whether to select between-chr-normalised columns}
}
\value{
data.table with normalised counts
}
\description{
Quantile normalize count data between features
}
