% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_functions.R
\name{plot_transform_effects}
\alias{plot_transform_effects}
\title{Plot the transformed data}
\usage{
plot_transform_effects(feature_dat, design, fn_stub, chr_sel = "chr11")
}
\arguments{
\item{feature_dat}{data.table with data columns to plot}

\item{design}{table with design information}

\item{fn_stub}{prepended to output file name}

\item{chr_sel}{example chromosome to plot}
}
\description{
Plot the transformed data
}
