% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varistran.R
\name{vst}
\alias{vst}
\title{Variance Stabilizing Transformation}
\usage{
vst(
  x,
  method = "anscombe.nb",
  lib.size = NULL,
  cpm = FALSE,
  dispersion = NULL,
  design = NULL
)
}
\arguments{
\item{x}{A matrix of counts. Rows are features, columns are samples.}

\item{method}{VST method; see details.}

\item{lib.size}{Optional library sizes.}

\item{cpm}{Return log2 CPM instead of simple log2.}

\item{dispersion}{Optional dispersion for negative binomial VST.}

\item{design}{Optional design matrix for dispersion estimation.}
}
\value{
Transformed matrix with attributes for method and library sizes.
}
\description{
Perform a Variance Stabilizing Transformation (VST) of a matrix of count data.
}
