% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varistran.R
\name{vst_advice}
\alias{vst_advice}
\title{Advise how VST will transform data}
\usage{
vst_advice(
  what = "anscombe.nb",
  dispersion = NULL,
  cpm = FALSE,
  lib.size = NULL
)
}
\arguments{
\item{what}{Either output of vst() or method name}

\item{dispersion}{As per vst()}

\item{cpm}{As per vst()}

\item{lib.size}{As per vst()}
}
\value{
Data frame showing transformed counts and twofold steps
}
\description{
Advise how VST will transform data
}
