% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_functions.R
\name{normalizeQuantile.betweenChr}
\alias{normalizeQuantile.betweenChr}
\title{Quantile normalize count data between chromosomes}
\usage{
\method{normalizeQuantile}{betweenChr}(
  feature_dat,
  rmap,
  chrs = paste0("chr", c(1:22, "X", "Y")),
  exclude_values = FALSE
)
}
\arguments{
\item{feature_dat}{data.table with columns fid and counts}

\item{rmap}{data.table with fid and chr}

\item{chrs}{vector of chromosomes to normalise between}

\item{exclude_values}{boolean or numeric vector length 2}
}
\value{
data.table with normalized counts
}
\description{
Quantile normalize count data between chromosomes
}
