% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_functions.R
\name{normalizeQuantile.betweenChr}
\alias{normalizeQuantile.betweenChr}
\title{Quantile normalize count data between chromosomes}
\usage{
\method{normalizeQuantile}{betweenChr}(
  countData,
  rmap,
  chrs = c(1:22, "X", "Y"),
  excludeVals = FALSE
)
}
\arguments{
\item{countData}{data.table with columns rfID and counts}

\item{rmap}{data.table with rfID and chr}

\item{chrs}{vector of chromosomes}

\item{excludeVals}{boolean or numeric vector length 2}
}
\value{
data.table with normalized counts
}
\description{
Quantile normalize count data between chromosomes
}
