% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_functions.R
\name{regressDistance}
\alias{regressDistance}
\title{Perform regression on score per restriction fragment}
\usage{
regressDistance(
  scorePerRF,
  rmap,
  regrType = "ols",
  isRef = FALSE,
  plotResids = FALSE,
  outDir = getwd(),
  exclude = FALSE,
  sel = NA
)
}
\arguments{
\item{scorePerRF}{data.table with rfID and counts}

\item{rmap}{data.table with rfID, chr, len}

\item{regrType}{regression type: "ols", "pois", "nb", or "olslog"}

\item{isRef}{logical, if TRUE label as reference}

\item{plotResids}{logical, whether to plot residuals}

\item{outDir}{output directory}

\item{exclude}{FALSE or numeric vector of length 2}

\item{sel}{numeric vector of row indices}
}
\value{
list with residualsPerRF and lm.summary
}
\description{
Perform regression on score per restriction fragment
}
