% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_functions.R
\name{regressDistance}
\alias{regressDistance}
\title{Perform regression on score per restriction fragment}
\usage{
regressDistance(
  score_per_frag,
  feature_ids,
  rmap,
  regrType = "ols",
  isRef = FALSE,
  plotResids = FALSE,
  outDir = getwd(),
  include = FALSE,
  sel = NA
)
}
\arguments{
\item{score_per_frag}{data.table with preprocessed counts per fragment}

\item{rmap}{data.table with rfID, chr, len}

\item{regrType}{regression type: "ols", "pois", "nb", or "olslog"}

\item{isRef}{logical, if TRUE label as reference}

\item{plotResids}{logical, whether to plot residuals}

\item{outDir}{output directory}

\item{include}{FALSE or numeric vector of length 2: include scores in interval}

\item{sel}{numeric vector of row indices}
}
\value{
list with residualsPerRF and lm.summary
}
\description{
Perform regression on score per restriction fragment
}
